LPS-76509 Move MBMessageIterator to module
