+++ modules/apps/collaboration/message-boards/message-boards-web/src/main/resources/META-INF/resources/message_boards/init.jsp
@@ -76,7 +76,6 @@ page import="com.liferay.message.boards.kernel.model.MBCategoryDisplay" %><%@
 page import="com.liferay.message.boards.kernel.model.MBMessage" %><%@
 page import="com.liferay.message.boards.kernel.model.MBMessageConstants" %><%@
 page import="com.liferay.message.boards.kernel.model.MBMessageDisplay" %><%@
-page import="com.liferay.message.boards.kernel.model.MBMessageIterator" %><%@
 page import="com.liferay.message.boards.kernel.model.MBThread" %><%@
 page import="com.liferay.message.boards.kernel.model.MBTreeWalker" %><%@
 page import="com.liferay.message.boards.kernel.service.MBCategoryLocalServiceUtil" %><%@
@@ -101,6 +100,7 @@ page import="com.liferay.message.boards.web.internal.display.context.MBDisplayCo
 page import="com.liferay.message.boards.web.internal.display.context.util.MBRequestHelper" %><%@
 page import="com.liferay.message.boards.web.internal.search.EntriesChecker" %><%@
 page import="com.liferay.message.boards.web.internal.util.MBBreadcrumbUtil" %><%@
+page import="com.liferay.message.boards.web.internal.util.MBMessageIterator" %><%@
 page import="com.liferay.message.boards.web.internal.util.MBSubscriptionUtil" %><%@
 page import="com.liferay.message.boards.web.internal.util.MBUserRankUtil" %><%@
 page import="com.liferay.message.boards.web.internal.util.MBWebComponentProvider" %><%@
