+++ portal-kernel/src/com/liferay/message/boards/kernel/model/MBMessageIterator.java
@@ -18,7 +18,9 @@ import java.util.Iterator;
 
 /**
  * @author Sergio González
+ * @deprecated As of 7.0.0, with no direct replacement
  */
+@Deprecated
 public interface MBMessageIterator extends Iterator<MBMessage> {
 
 	public int getIndexPage();
